Added a vtk_errors script to catch some of the vtk error output and print it rather than creating dialog windows. Seems to only effect certain versions. Also modified figure show to make it delete vtk components in the right order to avoid a vtk error on close (something about an invalid handle). Between these two changes all python versions I've tested no longer show the close error in a dialog which is a big improvement - the dialog windows are a pain. Python 3.5 still prints the error - no idea what it's problem is.
Added "from builins import super" all over the place to allow for
super().method(...) Python 3 style to work in Python 2.7. This adds
future as a dependency.

Added a minimal spell check (more like a case check) to the named color
lookup. So vpl.plot(..., color="Red") will be autocorrected to
color="red" and it'll print a message to say so. It checks case and
seperators i.e spaces rather than underscores.

Added a render_size property to the figures.

Realised the VTK's screenshot 'scale' parameter is relative to the
render size and not just output_size/300 where 300 is just some random
inexplainable constant as I origionally thought. Updated vpl.save_fig to reflect that.